Dummy implementation of solver and TreeDecomposition to check flow is working

+++ TreeVector.h
@@ -10,7 +10,7 @@
 using namespace std;
 
 struct TreeVector{
-    int runningTime;
+    float runningTime;
     vector<double> features;
 };
